requirements.txt
flask==2.3.3
ultralytics==8.0.196
opencv-python==4.8.1.78
pillow==10.0.1
requests==2.31.0
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
python-dotenv==1.0.0
huggingface_hub>=0.20.0
